+++ functions.php
@@ -136,8 +136,8 @@ add_action( 'widgets_init', '_s_widgets_init' );
 function _s_scripts() {
 	wp_enqueue_style( '_s-style', get_stylesheet_uri(), ['dashicons'] );
 
-	wp_enqueue_script( '_s-app', get_template_directory_uri() . '/js/app.js', [], false, true );
-	wp_enqueue_script( '_s-navigation', get_template_directory_uri() . '/js/navigation.js', [], false, true );
+	wp_enqueue_script( '_s-app', get_template_directory_uri() . '/js/app.js', ['behaviours'], false, true );
+	wp_enqueue_script( '_s-navigation', get_template_directory_uri() . '/js/navigation.js', ['behaviours'], false, true );
 
 	if ( is_singular() && comments_open() && get_option( 'thread_comments' ) ) {
 		wp_enqueue_script( 'comment-reply' );